Use JsonCodec for query params and responses
`jsonEncode()` takes care of adding quotes around strings, and
`jsonDecode()` returns strings without extra quotes and booleans as
booleans. This is essentially the same as what Subiquity does in Python
i.e. serializing and de-serializing to/from JSON. As a bonus, we get rid
of the silly dartx dependency for... removing quotes, tada! (: