Consistent naming of GET requests
Half of the GET requests had a get-prefix while the others (earliest
added ones) didn't. It's not only nice to have consistent but it also
emphasizes the behavior of the API. At the same time, this paves the
road for merging SubiquityStatusMonitor into SubiquityClient by freeing
up the "status" name for a synchronous getter for accessing the latest
known status. The end goal is to make it easier to access and present
the status anywhere in the GUI.