Import common Disk & Partition extensions from UDI (#104)
The sysname extensions are used everywhere in UDI because of being used
as display names for disks and partitions. Currently, they are exported
from the "disk storage service" which is being trimmed down and probably
split into smaller entities. Exporting them from here makes them flow
down everywhere SubiquityClient is used and where the storage service is
not necessarily needed.